+++ README.md
@@ -24,9 +24,3 @@ These projects can be found within GitHub parallel to this repository.
 
 - android-lib-appengine
 - android-lib-gameengine
-
-## Google Play Store
-
-The application is published on the following URL:
-
-https://play.google.com/store/apps/details?id=info.simplyapps.lwp.fishes
