+++ app/build.gradle
@@ -3,23 +3,23 @@ apply plugin: 'com.android.application'
 android {
     signingConfigs {
         debug {
-            storeFile file('N:\\Workspace\\keystore-debug')
+            storeFile file('..\\..\\keystore-debug')
             storePassword 'android'
         }
         release {
-            storeFile file('N:\\Workspace\\keystore-publish')
+            storeFile file('..\\..\\keystore-publish')
             storePassword 'keystore'
             keyAlias 'key'
             keyPassword 'keypass'
         }
     }
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdk 32
+    buildToolsVersion '32.1.0 rc1'
 
     defaultConfig {
-        applicationId "info.simplyapps.lwp.fishes"
-        minSdkVersion 23
-        targetSdkVersion 29
+        applicationId "com.juergenkleck.android.lwp.fishes"
+        minSdkVersion 24
+        targetSdkVersion 32
         versionCode = 6
         versionName = '1.2.0'
         signingConfig signingConfigs.release
@@ -41,8 +41,8 @@ android {
 }
 
 dependencies {
-    implementation "androidx.fragment:fragment:1.2.4"
-    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.fragment:fragment:1.4.1'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
     implementation project(path: ":appEngine")
     implementation project(path: ":gameEngine")
 }
