+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="info.simplyapps.lwp.fishes">
+    package="com.juergenkleck.android.lwp.fishes">
 
     <uses-feature android:name="android.software.live_wallpaper" />
 
@@ -13,7 +13,8 @@
         <service
             android:name=".engine.FishesLWP"
             android:label="@string/app_name"
-            android:permission="android.permission.BIND_WALLPAPER">
+            android:permission="android.permission.BIND_WALLPAPER"
+            android:exported="false">
             <intent-filter>
                 <action android:name="android.service.wallpaper.WallpaperService" />
             </intent-filter>
