+++ app/src/main/java/com/juergenkleck/android/lwp/fishes/engine/FishEngine.java
@@ -1,0 +1,12 @@
+package com.juergenkleck.android.lwp.fishes.engine;
+
+import com.juergenkleck.android.gameengine.engine.BasicEngine;
+
+/**
+ * Android app - FishesLWP
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public interface FishEngine extends BasicEngine {
+
+}
