+++ app/src/main/java/com/juergenkleck/android/lwp/fishes/preference/FishesEnum.java
@@ -1,0 +1,18 @@
+package com.juergenkleck.android.lwp.fishes.preference;
+
+/**
+ * Android app - FishesLWP
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public enum FishesEnum {
+
+    neon(0), bluetang(1), discus(2), cichlid(3);
+
+    public final int number;
+
+    private FishesEnum(int n) {
+        number = n;
+    }
+
+}
