+++ app/src/main/java/com/juergenkleck/android/lwp/fishes/preference/WallpaperEnum.java
@@ -1,0 +1,18 @@
+package com.juergenkleck.android.lwp.fishes.preference;
+
+/**
+ * Android app - FishesLWP
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public enum WallpaperEnum {
+
+    bluewater(0), reef1(1), reef2(2), bluewater2(3), reef3(4), rendered(5);
+
+    public final int number;
+
+    private WallpaperEnum(int n) {
+        number = n;
+    }
+
+}
