+++ app/src/main/java/com/juergenkleck/android/lwp/fishes/rendering/FishRendererTemplate.java
@@ -1,0 +1,24 @@
+package com.juergenkleck.android.lwp.fishes.rendering;
+
+import android.content.Context;
+
+import java.util.Properties;
+
+import com.juergenkleck.android.gameengine.rendering.GenericRendererTemplate;
+
+/**
+ * Android app - FishesLWP
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public abstract class FishRendererTemplate extends GenericRendererTemplate {
+
+    public FishRendererTemplate(Context context, Properties properties) {
+        super(context, properties);
+    }
+
+    public boolean logEnabled() {
+        return false;
+    }
+
+}
