+++ app/src/main/java/com/juergenkleck/android/lwp/fishes/rendering/RenderState.java
@@ -1,0 +1,12 @@
+package com.juergenkleck.android.lwp.fishes.rendering;
+
+/**
+ * Android app - FishesLWP
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public enum RenderState {
+
+    RUNNING, PAUSED;
+
+}
