+++ app/src/main/java/com/juergenkleck/android/lwp/fishes/storage/dto/Inventory.java
@@ -1,0 +1,20 @@
+package com.juergenkleck.android.lwp.fishes.storage.dto;
+
+import java.io.Serializable;
+
+import com.juergenkleck.android.appengine.storage.dto.BasicTable;
+
+/**
+ * Android app - FishesLWP
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public class Inventory extends BasicTable implements Serializable {
+
+    /**
+     *
+     */
+    private static final long serialVersionUID = -6974204243261183587L;
+    public boolean migrated;
+
+}
