--- app/src/main/java/info/simplyapps/lwp/fishes/engine/FishEngine.java
@@ -1,7 +1,0 @@
-package info.simplyapps.lwp.fishes.engine;
-
-import info.simplyapps.gameengine.engine.BasicEngine;
-
-public interface FishEngine extends BasicEngine {
-
-}
