--- app/src/main/java/info/simplyapps/lwp/fishes/preference/FishesEnum.java
@@ -1,13 +1,0 @@
-package info.simplyapps.lwp.fishes.preference;
-
-public enum FishesEnum {
-
-    neon(0), bluetang(1), discus(2), cichlid(3);
-
-    public final int number;
-
-    private FishesEnum(int n) {
-        number = n;
-    }
-
-}
