--- app/src/main/java/info/simplyapps/lwp/fishes/preference/ValueContainer.java
@@ -1,18 +1,0 @@
-package info.simplyapps.lwp.fishes.preference;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ValueContainer extends info.simplyapps.gameengine.rendering.data.ValueContainer {
-
-    public String wallpaper;
-    public Map<String, Boolean> fishes;
-    public boolean swimSchool;
-    public boolean swimRealistic;
-    public int swimSpeed;
-
-    public ValueContainer() {
-        fishes = new HashMap<String, Boolean>();
-    }
-
-}
