--- app/src/main/java/info/simplyapps/lwp/fishes/preference/WallpaperEnum.java
@@ -1,13 +1,0 @@
-package info.simplyapps.lwp.fishes.preference;
-
-public enum WallpaperEnum {
-
-    bluewater(0), reef1(1), reef2(2), bluewater2(3), reef3(4), rendered(5);
-
-    public final int number;
-
-    private WallpaperEnum(int n) {
-        number = n;
-    }
-
-}
