--- app/src/main/java/info/simplyapps/lwp/fishes/rendering/FishRendererTemplate.java
@@ -1,19 +1,0 @@
-package info.simplyapps.lwp.fishes.rendering;
-
-import android.content.Context;
-
-import java.util.Properties;
-
-import info.simplyapps.gameengine.rendering.GenericRendererTemplate;
-
-public abstract class FishRendererTemplate extends GenericRendererTemplate {
-
-    public FishRendererTemplate(Context context, Properties properties) {
-        super(context, properties);
-    }
-
-    public boolean logEnabled() {
-        return false;
-    }
-
-}
