--- app/src/main/java/info/simplyapps/lwp/fishes/rendering/RenderState.java
@@ -1,7 +1,0 @@
-package info.simplyapps.lwp.fishes.rendering;
-
-public enum RenderState {
-
-    RUNNING, PAUSED;
-
-}
