--- app/src/main/java/info/simplyapps/lwp/fishes/rendering/objects/FishObject.java
@@ -1,22 +1,0 @@
-package info.simplyapps.lwp.fishes.rendering.objects;
-
-
-public class FishObject {
-
-    // the click event
-    public boolean init;
-    // what happens on click
-    public int clickAction;
-
-    // the view angle in the 3d space
-    // needed for rotation and transformation
-//	public float angle;
-    // the rotation of the object
-//	public float rotation;
-
-    public int[] gReferences;
-
-    public FishObject() {
-    }
-
-}
