--- app/src/main/java/info/simplyapps/lwp/fishes/storage/dto/Inventory.java
@@ -1,15 +1,0 @@
-package info.simplyapps.lwp.fishes.storage.dto;
-
-import java.io.Serializable;
-
-import info.simplyapps.appengine.storage.dto.BasicTable;
-
-public class Inventory extends BasicTable implements Serializable {
-
-    /**
-     *
-     */
-    private static final long serialVersionUID = -6974204243261183587L;
-    public boolean migrated;
-
-}
