+++ build.gradle
@@ -8,7 +8,7 @@ buildscript {
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.2'
+        classpath 'com.android.tools.build:gradle:7.3.0-alpha01'
         
 
         // NOTE: Do not place your application dependencies here; they belong
